refactor(database): Cursor should be int64

+++ internal/database/cursor.go
@@ -3,14 +3,14 @@ package database
 import "github.com/samber/lo"
 
 type Cursor struct {
-	Offset int
-	Limit  int
+	Offset int64
+	Limit  int64
 }
 
-func (c Cursor) GetOffset() int {
+func (c Cursor) GetOffset() int64 {
 	return c.Offset
 }
 
-func (c Cursor) GetLimit() int {
+func (c Cursor) GetLimit() int64 {
 	return min(lo.CoalesceOrEmpty(c.Limit, 10), 100)
 }
